# src/main/java/com/example/springboot4demo/pricing/dto/PriceEvent.java
package com.example.springboot4demo.pricing.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class PriceEvent {
    @JsonProperty("product_sku")
    private String productSku;

    @JsonProperty("product_id")
    private Long productId;

    @JsonProperty("amount")
    private BigDecimal amount;

    @JsonProperty("currency")
    private String currency;

    @JsonProperty("source")
    private String source;

    @JsonProperty("timestamp")
    private LocalDateTime timestamp;
}